operator: Refactor class hierarchy [WIP]
